Add altair LayerChart to schema (#194)
**What does this PR suggest?**
Adding two different altair Charts creates a LayerChart, which stacks
figures. Since this is a basic altair feature and
we currently only support Chart, LayerChart would be a useful addition
for https://github.com/probabl-ai/mandr/pull/186. I didn't add an issue
for this PR as this is a quick fix.

WDYT?