Manually trigger a release on a commit in the GitHub UI (#267)
The creation of a release can be summary to:

1. building the frontend and sharing the library to mandr,
2. building the mandr package distributions using the tag name as
version,
3. publishing to PyPI.

---

Release workflow is triggered when creating release using the [Github
UI](https://github.com/probabl-ai/mandr/releases/new).
Release can actually only be based on release tag or release-candidate
tag from the main branch using Github
[environment](https://docs.github.com/en/actions/managing-workflow-runs-and-deployments/managing-deployments/managing-environments-for-deployment).
Release is done using [trusted
publisher](https://docs.pypi.org/trusted-publishers/) for token
isolation questions.

Tags to release follow the semantic versioning with a "v" prefix (for
instance, `v1.0.0` or `v1.0.0-rc.1`).
Tags to release can only be created by repository admins using Github
[rules](https://docs.github.com/en/enterprise-server@3.14/repositories/managing-your-repositorys-settings-and-features/managing-repository-settings/configuring-tag-protection-rules).

Release version is **_only_** based on the tag name. Developer is
responsible for it. Keep in mind that it will shadow the (actually
untrustable) version specified in `VERSION.txt` or `pyproject.toml`.

This PR is a first attempt to create a release workflow, but is not
complete. It must be complete in a near future with strong and automatic
version definition, automatic change-log creation and security
restrictions enhancement.

Notes: Jobs are isolated due to the security issues on
[permissions](https://docs.pypi.org/trusted-publishers/security-model/).
Publishing to PyPI using trusted publisher has consequences.