feat(UI): Implement a `DraggableList` component (#460)
This PR is the first step to fix #408.

## UI/UX preview


https://github.com/user-attachments/assets/d9db0c12-bee1-4795-ab3b-fc7c4e11ae9e



## usage

The component needs a reactive list of `Item` (a ref). Item is a regular
object that needs an id filed that is a string.
It will mutate the given array of items to reflect the user selected
order.

```vue
<script setup lang="ts">
import { ref } from "vue";

const draggableListData = ref(
  Array.from({ length: 25 }, (v, i) => ({
    id: `${i}`,
    color: `hsl(${(360 / 25) * i}deg, 90%, 50%)`,
    content: Array.from(
      { length: Math.floor(Math.random() * 10) + 1 }, // Random number of items between 1 and 10
      () => String.fromCharCode(97 + Math.floor(Math.random() * 26)) // Random lowercase letter
    ),
  }))
);
</script>

<template>
  <DraggableList v-model:items="draggableListData">
    <template #item="{ id, color, content }">
      <div :style="{ backgroundColor: color, color: 'white' }">
        <span>ID: {{ id }}</span>
        <ul>
          <li v-for="(c, i) in content" :key="i">{{ c }}</li>
        </ul>
      </div>
    </template>
  </DraggableList>
</template>
```

## implementation choices

The implementation introduces two new dependencies. 
- [html-to-image](https://github.com/bubkoo/html-to-image) to create
rasterized version of the moved item
- [interact.js](https://interactjs.io/) to streamline pointer
interaction

The main container is a flexbox, items flow normally in it (no absolute
positioning).
When a drag start the selected items becomes slightly transparent and a
raster copy of it is generated. That raster copy follows user's pointer
using computed translation (CSS transform). To make room for the item to
be dropped each items also has a drop indicator. The size of this drop
indicator reflects the size of the dragged item. I use margin top &
bottom to give different feedback when dragging down or up. This is nice
because no coordinates calculation is needed and items continue to flow
normally in the container.

To limit clashes with inner item code when dragging all the items are
transformed to a rasterized version of themself.

## known limitation

~~For now it is not possible to automatically scroll the container when
dragging outside of its visible range.~~

---------

Co-authored-by: Thomas S. <thomas@probabl.ai>