fix(docs): Fix example
Currently cross-validation aggregation works by manipulating the contents of the keys "cross_validation" and "cross_validation_aggregated", and we assume that the user is only working on one model so that it makes sense to compare different runs in a plot.

In example 3 we first perform cross-validation on a classification model, and then later on a regression model. Because of this, the aggregation process fails and the docs cannot be built.

This is a dirty, quick fix where we delete the contents of the keys "cross_validation" and "cross_validation_aggregated" before the regression example, to start fresh.

We are already planning more robust, automatic solutions to relax the "one-model" assumption: see
- https://github.com/probabl-ai/skore/issues/557
- https://github.com/probabl-ai/skore/issues/558