refactor(CrossValidationItem): Move default parameters detection from route to CrossValidationItem (#970)
As requested in
https://github.com/probabl-ai/skore/pull/871#discussion_r1880574108

In the UI, in the widget for `CrossValidationItem`, we display the
estimator's parameters as well as whether each parameter is the default
for that estimator. This information was computed at the "send to UI"
step of the process, but we now compute it just once, when the
`CrossValidationItem` is created (i.e. at `put` time).

In addition, we now no longer throw in `put` if the tested estimator has
no `get_params` method.

---------

Co-authored-by: Thomas S. <thomas@probabl.ai>