feat: Use rich Panel for showing warning in train_test_split (#1086)
closes #1060 

It is the alternative to #1060 using `rich`. I added a test to check
that we can filter the warning since we are not using the usual
`warnings` module.

In the future, we could factor out the code in a utils to be sure that
we can also transform the warnings into error.