feat: Add scoring_names to overwrite the column of the score dataframe (#1114)
closes #1111 

Add a `scoring_names` parameter allowing to overwrite the columns of
dataframe returned by the `report_metrics`.

Build upon #1109 that expose some private functions for our internal
metrics. It is nice because we can therefore expose as well the
`metric_name` consistently in both `_custom_metric` and our defined
metric.