fix: Fix rendering in ipywidget by injecting CSS rule asap (#1139)
closes #1132

The current patch was clean but not optimal: it was patching a public
class. However, it involves calling `display()` once to create the
ipywidget and then once for injection the CSS rule. I assume that the
timing between the two displays call was actually the blinking observe.

This PR patch the function creating the HTML code block for the widget
and concatenate the CSS rule and thus involve a single call to display
and remove the blinking.