feat(UI): Add markdown editor to annotate item (#1082)
Part of #1041 

- [x] upgrade dependencies (eslint 9 \o/)
- [x] new icons
- [x] minimal skinnable rich text editor
- [x] expose python's `set_note` as a JSON API 
- [x] plug the API in the frontend

## UI preview


https://github.com/user-attachments/assets/33cdf005-9ddd-42db-8059-c84c6a4a2433

## implementation choices

Finding a good and not too opinionated rich text editor is no easy job.
After struggling with a few libs I decided that rolling a minimal editor
may do the job (at least for now).

The RichTextEditor component (probably badly named) tries to:
- help user with basic markdown syntax
- keep interactions button outside of it for best skinnability

In a future work we may move to something more powerful.