fix: Make reports pickable out-of-the-box (#1179)
This PR makes sure that the `EstimatorReport` and
`CrossValidationReport` are pickable object.

Right now, because we kept the step of the `Progress` from rich that has
a lock, the object are not pickable. I reset the `_parent_progress` and
`_progress_info` to `None` in the finally of the decorator.

An alternative is to change the `set_state` but here I think that we can
just applied this clean-up instead.