docs: Add example to showcase EstimatorReport and CrossValidationReport (#1156)
Build upon https://github.com/probabl-ai/skore/pull/1091
We should merge https://github.com/probabl-ai/skore/pull/1091 to make it
easy to review.

This PR revisits the documentation for the `EstimatorReport` and
`CrossValidationReport`.

The idea is to provide different examples for different purposes:

- technical section: it is a section to explain more in details some
internal for user that might be interested. I think that we should move
those inside this specific section because we should have more a pure
data science section to only show the added values of our tools on data
science
- use-case section: it is indeed the section to show case our tools on
real data. Here we want to show the added value on the data science
side.

---------

Co-authored-by: Sylvain Combettes <48064216+sylvaincom@users.noreply.github.com>
Co-authored-by: Auguste Baum <52001167+augustebaum@users.noreply.github.com>
Co-authored-by: Matt J. <matthieu@probabl.ai>
Co-authored-by: Thomas S. <th.salvatore@gmail.com>
Co-authored-by: MarieS-WiMLDS <79304610+MarieS-WiMLDS@users.noreply.github.com>