feat(_find_ml_task): Refine ML task detection logic (#1305)
ML detection now also works for multioutput targets (e.g. where `y` is a
2d array).

Added 3 new MLTask variants: `multioutput-binary-classification`,
`multioutput-multiclass-classification` and `multioutput-regression`.
`regression` still means "single-output regression".

The detection uses the same mechanism described in
https://github.com/probabl-ai/skore/commit/be29a2641839e6ae83831ca1210b278d7f515c2a
to discriminate between an array of integers that is actually for
classification vs. one that is for regression.

Closes #1005