fix: Do not use aspect ratio of 1 for residual plot (#1335)
closes #1292 

We forced the aspect ratio of 1.0 for the residual plot. However the
residual is computing `actual vs predicted` and the scale of the
residual can be really different from the prediction if the error is
low, explaining the problem of the plot.

This PR only set the aspect ratio for the actual vs predicted plot only.