docs: Fix the ambuiguity regarding global configuration (#1349)
This PR tries to disambiguate what we mean by "global" when it comes to
the configuration setting.

In short global refer that we are going to dispatch the configuration to
local threads or sub-process.

It is really different from the "local" meaning within the code impacted
by the context manager.