feat(repr): Give a description to attributes (#1366)
Closes #1235   

I had to go through regex because from what I understand from inspect,
there is no way to access the documentation of an attribute. There is
very little risk that the regex matches something before the attribute
description because of naming overlap. If we think that this risk is to
big, as there is no way to do a test, we should close #1235 as not
possible.

What it does for an EstimatorReport:  

![image](https://github.com/user-attachments/assets/66da322a-dcb7-4d69-abdd-432076be6906)


What it does for a CrossValidationReport:  

![image](https://github.com/user-attachments/assets/fbd51eb8-5061-4bee-b3af-7b43d57eb65b)


What it does for a ComparisonReport:  

![image](https://github.com/user-attachments/assets/f3989da9-d1d6-4561-a5bf-95310fe04648)