docs: Fix rendering using custom matplotlib scraper (#1516)
It looks that the `display.figure_.tight_layout()` does not work for the
purpose of the documentation. I was looking within notebooks and VS code
but forgot to check the website.

For the website, we can actually solve it by modifying the matplotlib
scrappers from sphinx-gallery and fix pass the right parameter to
`savefig`. I removed the "200 dpi" because it makes the figure too big
and I don't think that the default quality is bad.