Make it impossible to pass the same CVReport several times
Now that progress info is only held by Reports, not accessors, if a
ComparisonReport of CrossValidationReports holds the same
CrossValidationReport several times, when `self_obj`'s progress
information is cleaned up, it will be cleaned up for the references as
well, which will lead to bugs.

Since it does not make sense to compare the exact some Report to itself,
a simple solution is to make it forbidden to pass the same Report more
than once.