refactor(progress_decorator): Make sure `self_obj` is a `Report` (#1536)
This originates from a bug when implementing comparison of
CrossValidationReport in #1512

Because CrossValidationReports hold a `_parent_progress`, when a
ComparisonReport creates a progress bar to iterate over
CrossValidationReports, the outer progress bar conflicts with the inner
progress bars, and rich refuses to proceed.

The solution is for the ComparisonReport to explicitly set its inner
CrossValidationReports' progress instance, so that in total there is
only one progress instance.

But before this change, the progress instance was sometimes owned by a
`CrossValidationReport.metrics` accessor. This is a problem because
accessors are re-instantiated whenever they are accessed, so their state
cannot be modified from the parent.

The solution this change implements is to remove all `progress`-related
attributes from all accessors, and to ensure that the progress instance
is only owned by the Report object, not by any of its accessors.