docs: Pass format to `pd.to_datetime` (#1542)
Closes #1504.

I simply inspected the `X_employee["date_first_hired"]` series by
printing it, and saw that it uses the `mm/dd/yyyy` format. I have
updated both examples to use the corresponding format argument.

The linked issue also mentions `skrub.ToDatetime` but I am not clear on
what exactly is to be done with that. It looks like that API is only
used in one example: `examples/use_cases/plot_employee_salaries.py`. If
anyone can clarify what needs to be done I would love to take it up in
this PR or a future PR.