ci: Introduce `<backend|sphinx>` jobs as "all green" required jobs (#1514)
Some changes to prepare the setup of the GH merge queue, for which we
need to specify the jobs required to succeed:

- Set filenames as workflow/job names to avoid confusion,
- Add `backend` job that can be used as required 🟢 job,
- Add `sphinx` job that can be used as required 🟢 job,
- Add explicit restricted permissions to the `sphinx` workflow (no
worries, they were already restricted to read-only).

---

For a job to be required, its **workflow** has to be executed and not
skipped.
Before this PR, we skipped entire workflow via path filtering, which is
incompatible with "required jobs".


https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks


![image](https://github.com/user-attachments/assets/a48bab3f-c361-4a55-b933-9cf94057f26c)

Now we always execute workflows, but skip jobs individually based on
modified files.