docs: Avoid failure when pytorch load weights from TextEncoder with parallelism (#1617)
It is a workaround on the failure that is sometimes happening when
request `transform` from `TextEncoder` in parallel.

In short, there is a failure in the weights loading from PyTorch where
some tensors are loaded on a meta device (only metadata) and does not
load the weight. I did not yet found the root cause of the failure why
the weights are not loaded properly.

In the meanwhile, we can avoid this issue by storing the weights. We
should monitor that loading several time the language model does not
make blow up the RAM on the GitHub actions. Othwerwise, we need to
deactivate the parallelism most probably.

Co-authored-by: Thomas S. <thomas@probabl.ai>