fix(Comparison[CVReport]): Avoid mutating DataFrames in cache (#1711)
Closes #1706
Closes #1704

The issue is a consequence of pandas DataFrame mutability.
Specifically, `_combine_cross_validation_results` manipulates
a list of DataFrames mutably, and since the DataFrames come
from a cache (which is a mutable dict), the DataFrames inside
the cache end up mutated.


https://github.com/probabl-ai/skore/blob/9bbfdb34fc8a62555aa8d3cc79a8f914dd22dd43/skore/src/skore/sklearn/_comparison/metrics_accessor.py#L250

`_combine_cross_validation_results` already performs a copy
of its input (a list of DataFrames) to prevent this sort of problem,
but apparently this is not enough: the copied list still references
the same DataFrames.


https://github.com/probabl-ai/skore/blob/9bbfdb34fc8a62555aa8d3cc79a8f914dd22dd43/skore/src/skore/sklearn/_comparison/utils.py#L295

---------

Co-authored-by: Guillaume Lemaitre <guillaume@probabl.ai>