feat!(plots): Remove the ability to pass `ax`
Starting from https://github.com/probabl-ai/skore/pull/1669, it
is sometimes possible to draw plots onto (a `ndarray` of) several
Axes rather than just a single Axes. In fact, starting from
https://github.com/probabl-ai/skore/pull/1720, the same data might be
drawn on one or the other depending on user parameters.

In this context, allowing the user to pass `ax` might add too much
complexity compared to how useful it might be; it is considered an
advanced use-case which is difficult to motivate.

This PR thus removes the ability to pass user-created `ax`.