docs: Document the skore.project.metadata.Metadata class (#1754)
Right now, we don't document `skore.project.metadata.Metadata`.

In short, it means that no one is aware of the API of the output of
`project.reports.metadata()` and what to do with it.

I avoid to list the inherited methods from the pandas dataframe since
the import in this case would not work. Instead I modify the
documentation to refer to the original documentation in the docstring.