ci: Fix `pytest.yml` on pipeline triggered by a pull request from a fork (#1766)
Fixing
https://github.com/probabl-ai/skore/actions/runs/15254302350/job/42898127740?pr=1669.

`vars` are not passed to workflows that are triggered by a pull request
from a fork.

This PR fixes this issue, by defining the packages list directly in the
workflow.

### Note
Environment variables are only available at the steps level, not in the
matrix. This is why the matrix uses the output of its need.

### Source

![image](https://github.com/user-attachments/assets/ad86c09f-e7ca-45ff-870c-9a4bc29ac12d)

![image](https://github.com/user-attachments/assets/eb13a9f1-002b-4a3e-acea-eb3748ef89bf)