refactor: Make the same variable common (#1826)
I have to use this variable in another PR
(https://github.com/probabl-ai/skore/pull/1816), and I find it not clean
to have to copy/paste the same information in different classes.

Refactoring this variable + the methods related to the helpers.  
More could be done having a look at the metrics and the displays, but no
real value for now. Let's wait for it to be needed.