ci: Fix pytest coverage reports to be interactive (#1834)
The `pytest.yml` workflow installs packages in a `venv` to run in real
condition.
It computes the coverage from it, instead of using directly the source. 

It has the impact of writing soft-link to the `venv` in the coverage
report, making the report non-interactive.
This PR solves this problem by replacing occurences of `venv` in the
report by the source paths.