chore(deps): Bump urllib3 in lockfiles (#1852)
Following https://github.com/probabl-ai/skore/pull/1850, and to close
dependabot security issues.

Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.10/scikit-learn-1.4
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.10/scikit-learn-1.7
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.11/scikit-learn-1.4
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.11/scikit-learn-1.7
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.12/scikit-learn-1.7
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.13/scikit-learn-1.5
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.13/scikit-learn-1.6
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore-hub-project/python-3.13/scikit-learn-1.7
directory: [urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.10/scikit-learn-1.4 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.10/scikit-learn-1.7 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.11/scikit-learn-1.4 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.11/scikit-learn-1.7 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.12/scikit-learn-1.4 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.12/scikit-learn-1.7 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.13/scikit-learn-1.5 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.13/scikit-learn-1.6 directory:
[urllib3](https://github.com/urllib3/urllib3).
Bumps the pip group with 1 update in the
/ci/requirements/skore/python-3.13/scikit-learn-1.7 directory:
[urllib3](https://github.com/urllib3/urllib3).

Updates urllib3 from 2.4.0 to 2.5.0.