feat(skore-hub-project)!: Send pickled report to S3 (#1891)
Closes https://github.com/probabl-ai/skore/issues/1878.

This PR changes the way reports are sent to the hub and retrieved from
it, with as few changes as possible.

A complete refactor will be done in a future iteration: we don't need
the "item" notion anymore, as long as the only supported objects are
`skore` reports. It can simplify a lot the code of `skore-hub-project`.