docs: Add tf-keras dependency
This resolves an issue with using `skrub.TextEncoder`.

The model created line 93 of plot_employee_salaries.py uses TextEncoder,
which fails when working on column "division":

```
>>> TextEncoder().fit_transform(df["division"])
... [tensorflow logs]
*** RuntimeError: Failed to import transformers.integrations.integration_utils because of the following error (look up to see its traceback):
Failed to import transformers.modeling_tf_utils because of the following error (look up to see its traceback):
Your currently installed version of Keras is Keras 3, but this is not yet supported in Transformers. Please install the backwards-compatible tf-keras package with `pip install tf-keras`.
```

Following the recommendation resolves the issue.