chore(skore): Remove find_estimators utility (#1994)
This PR simplifies https://github.com/probabl-ai/skore/pull/1965 and
allow to get a clean history for the `find_estimators` functionality.

For the sake of completness, we remove this utility after dropping the
feature discussed in https://github.com/probabl-ai/skore/issues/1811
because it was not obvious how to add every components of a pipeline on
the parallel coordinate plot.