fix(skore): Don't rely on _repr_html_ to display the summary widget (#2059)
Fixes #2039

The colab button to convert a dataframe into an interactive table
appears to call the dataframe's `_repr_html_`, so when it is used on a
summary (which inherits from the dataframe class), the widget would be
displayed a second time as `_repr_html_` is called again.

The proposed solution is to display the widget in `_repr_mimebundle_`
and override `_repr_html_` to return an empty string.