feat(skore): Improve layout using plt.tight_layout by default (#2182)
This PR is factor out from #2152 

The `plt.tight_layout` is a good way to not get issue with labels on the
y- / x-axis that could overlap. This PR is making call this function in
the decorator `style_plot` that is always call by the display when
calling `plt.tight_layout`.